# docs/swagger.yaml
basePath: /
definitions:
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  model.Subscription:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.TotalResponse:
    properties:
      count:
        type: integer
      total_cost:
        type: integer
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST‑сервис для учёта онлайн‑подписок пользователей.
  title: Subscriptions Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Получить список подписок с возможностью фильтрации по пользователю
        и сервису
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Добавляет новую запись о подписке
      parameters:
      - description: Subscription details
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Создать новую подписку
      tags:
      - subscriptions
  /subscriptions/{user_id}/{service_name}:
    delete:
      description: Удаление подписки по ID пользователя и имени сервиса
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Наименование сервиса
        in: path
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Получить подписку по ID пользователя и имени сервиса
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Имя сервиса
        in: path
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Изменение данных подписки по ID пользователя и имени сервиса
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Наименование сервиса
        in: path
        name: service_name
        required: true
        type: string
      - description: Измененные данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Изменить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Получите общую стоимость и количество подписок за определенный
        период с дополнительной фильтрацией
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Наименование сервиса
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: start_period
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: end_period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TotalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получить общую стоимость подписок
      tags:
      - subscriptions
swagger: "2.0"